Update and rename SyndicateFrameNFT.sol to DoNotUse.sol
This contract does not protect against erraneous minters, there is no ability to limit the amount of NFTs each wallet can obtain. 
This contract is also missing the list of authorized agents as referenced in the original template which are important for security. There are also no royalties applied or enforced.